04.26

+++ Jungol/src/q512/Main.java
@@ -4,7 +4,8 @@ public class Main {
 
 	public static void main(String[] args) {
 		int weight = 49;
-		double G = 0.2683;
+		float G = 0.2683F;
+		
 		
 		System.out.print("49 * 0.268300 = " + weight * G);
 		// 나는 소수점 자리수 늘리는 법을 몰라요
